Объединил методы в один файл
